requirements.txt
aioredis==1.2.0
appdirs==1.4.3
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
autopep8==1.4.4
Babel==2.6.0
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.6.1
daphne==2.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.39.1
django-ckeditor==5.7.1
django-cleanup==3.2.0
django-cors-headers==2.5.3
django-countries==5.3.3
django-el-pagination==3.2.4
django-js-asset==1.2.2
django-social-share==1.3.2
django-widget-tweaks==1.4.3
djangorestframework==3.9.3
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isodate==0.6.0
lxml==4.3.4
msgpack==0.6.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.0.0
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
stripe==2.32.1
toml==0.10.0
Twisted==19.2.0
txaio==18.8.1
urllib3==1.24.3
Werkzeug==0.15.2
whitenoise==4.1.2
zeep==3.4.0
zope.interface==4.6.0
